+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Version Unreleased ()
 --------------------------------
 
+* Added: get_files to list files in a given torrent
+
 * Bugfix: Problem with start/stop of deluge 1 torrents
 
 Version 1.0.0 (10-05-2020)